1.抽象方法文件名前加上Abstract 2.第三方信息infoprovider返回时都调用一下return StructSerializer.toString(obj);去掉thrift中的多余项

+++ position-service-provider/src/main/java/com/moseeker/position/service/position/base/refresh/handler/AbstractRegionResultHandler.java
@@ -12,15 +12,14 @@ import org.apache.commons.lang.time.FastDateFormat;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 import java.util.Locale;
 
-public abstract class RegionResultHandler extends AbstractJsonResultHandler {
-    Logger logger= LoggerFactory.getLogger(RegionResultHandler.class);
+public abstract class AbstractRegionResultHandler extends AbstractJsonResultHandler {
+    Logger logger= LoggerFactory.getLogger(AbstractRegionResultHandler.class);
 
     @Autowired
     DictCityMapDao cityMapDao;
@@ -35,40 +34,45 @@ public abstract class RegionResultHandler extends AbstractJsonResultHandler {
 
         //处理不匹配的数据
         List<Region> moseekerUnmatched=regionMap.getJSONArray("moseekerUnmatched").toJavaList(Region.class);
-        logger.info("moseekerUnmatched:{}",moseekerUnmatched);
+        logger.info("moseekerUnmatched size:{}",moseekerUnmatched==null? "null":moseekerUnmatched.size());
         List<Region> thirdPartyUnmatched=regionMap.getJSONArray("thirdPartyUnmatched").toJavaList(Region.class);
-        logger.info("thirdPartyUnmatched:{}",thirdPartyUnmatched);
+        logger.info("thirdPartyUnmatched size:{}",thirdPartyUnmatched==null? "null":thirdPartyUnmatched.size());
 
         //处理匹配的数据
         JSONArray matchedDatas=regionMap.getJSONArray("map");
-        logger.info("matchedDatas:{}",matchedDatas);
+        if(matchedDatas==null) {
+            logger.info("matchedDatas:{}", "null");
+        }else{
+            List<DictCityMapDO> cityMapList=new ArrayList<>();
+            String nowTime= FastDateFormat.getDateInstance(FastDateFormat.LONG, Locale.CHINA).format(new Date());
+            for(int i=0;i<matchedDatas.size();i++){
+                JSONObject data=matchedDatas.getJSONObject(i);
+                Region moseekerRegion=data.getJSONObject("moseekerRegion").toJavaObject(Region.class);
+                Region thirdPartyRegion=data.getJSONObject("thirdPartyRegion").toJavaObject(Region.class);
 
-        List<DictCityMapDO> cityMapList=new ArrayList<>();
-        String nowTime= FastDateFormat.getDateInstance(FastDateFormat.LONG, Locale.CHINA).format(new Date());
-        for(int i=0;i<matchedDatas.size();i++){
-            JSONObject data=matchedDatas.getJSONObject(i);
-            Region moseekerRegion=data.getJSONObject("moseekerRegion").toJavaObject(Region.class);
-            Region thirdPartyRegion=data.getJSONObject("thirdPartyRegion").toJavaObject(Region.class);
+                int moseekerCode= PositionRefreshUtils.lastCode(moseekerRegion.getCode());
 
-            int moseekerCode= PositionRefreshUtils.lastCode(moseekerRegion.getCode());
+                DictCityMapDO cityMap=new DictCityMapDO();
 
-            DictCityMapDO cityMap=new DictCityMapDO();
+                cityMap.setCreateTime(nowTime);
+                cityMap.setCodeOther(thirdPartyRegion.codeToString());
+                cityMap.setStatus(0);
+                cityMap.setCode(moseekerCode);
+                cityMap.setChannel(channelValue);
 
-            cityMap.setCreateTime(nowTime);
-            cityMap.setCodeOther(thirdPartyRegion.codeToString());
-            cityMap.setStatus(0);
-            cityMap.setCode(moseekerCode);
-            cityMap.setChannel(channelValue);
+                cityMapList.add(cityMap);
+            }
 
-            cityMapList.add(cityMap);
+            logger.info("region for insert : {}",cityMapList);
+
+            int delCount=cityMapDao.delete(new Condition("channel",channelValue));
+            logger.info("channel {} delete old region ",channelValue,delCount);
+            cityMapDao.addAllData(cityMapList);
+            logger.info("channel {} region insert success",channelValue);
         }
 
-        logger.info("region for insert : {}",cityMapList);
 
-        int delCount=cityMapDao.delete(new Condition("channel",channelValue));
-        logger.info("channel {} delete old region ",channelValue,delCount);
-        cityMapDao.addAllData(cityMapList);
-        logger.info("channel {} region insert success",channelValue);
+
     }
 
     private static class Region {
